refactor(api): dedupe DO-namespace fake + route-name the tracker params schema (review #469)
- Add aFakeDurableObjectNamespaceWith to base do.fake and use it for both DO
  bindings in aFakeEnvWith (which already provides the default INDIVIDUAL_TRACKER_DO
  stub) — removes the duplicated namespace boilerplate. Tests override the stub via
  aFakeEnvWith({ INDIVIDUAL_TRACKER_DO: aFakeDurableObjectNamespaceWith(stub) })
  instead of a per-test envWithTrackerDo helper.
- Rename trackerIdParamsSchema -> trackerParamsSchema (route-oriented params bag,
  not param-specific).
