+++ shared/src/contracts/individual-tracker/tracker.ts
@@ -39,10 +39,10 @@ export const selectActiveTrackerRequestSchema = z.object({
 });
 export type SelectActiveTrackerRequest = z.infer<typeof selectActiveTrackerRequestSchema>;
 
-export const trackerIdParamsSchema = z.object({
+export const trackerParamsSchema = z.object({
   trackerId: z.string().min(1),
 });
-export type TrackerIdParams = z.infer<typeof trackerIdParamsSchema>;
+export type TrackerParams = z.infer<typeof trackerParamsSchema>;
 
 export const trackerContract = defineContract(z.object({ tracker: trackerSchema }));
 export type TrackerResponse = z.infer<typeof trackerContract.schema>;
